Custom emojis + reactions + custom emoji reactions
There is a bug I'll fix tomorrow related to adding 2 normal emojis as a reaction to a message with an existing custom emoji reaction, the 2 normal emojis will merge together in count and become 1 emoji with like invisible reactors. lol. I think It's something with the null ID the default emojis use for reactions.